refactor: Extract types + hooks for Contracts, SuperAdmin, StrategieRubata, Trades
Extract state/logic into dedicated hooks and type files:
- Contracts.tsx: 2,434 → 1,675 lines (-31%)
- SuperAdmin.tsx: 2,402 → 1,792 lines (-25%)
- StrategieRubata.tsx: 2,091 → 1,456 lines (-30%)
- Trades.tsx: 2,069 → 1,485 lines (-28%)

Total: -2,922 lines removed from components, moved to 4 type files + 4 hook files
